+++ storybook/stories/LanguageSelectionForm.stories.js
@@ -25,6 +25,7 @@ storiesOf('LanguageSelectionForm', module)
       <LanguageSelectionForm
         onSubmit={action('submit')}
         languages={LANGUAGES}
+        preselectedLanguage={LANGUAGES[0].value}
         isSubmitting={false}
       />
     </div>
@@ -35,6 +36,7 @@ storiesOf('LanguageSelectionForm', module)
       <LanguageSelectionForm
         onSubmit={action('submit')}
         languages={LANGUAGES}
+        preselectedLanguage={LANGUAGES[0].value}
         isSubmitting
       />
     </div>